Remove deprecated chat-synapse step definitions and consolidate login functionality into login.steps.mjs. Enhance E2E test coverage for chat interactions, including message visibility, image previews, and reply handling.
